# openapi.yaml
components:
  schemas:
...
    http.errorResponse:
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
      type: object
info:
  contact:
    email: info@alkem.io
    name: Alkemio Foundation
    url: https://alkem.io
  description: |
    Universal file I/O gateway and document lifecycle manager for the Alkemio platform.
    Provides authenticated file serving (public), document CRUD, image processing,
    and MIME validation (internal).
  license:
    name: EUPL-1.2
    url: https://opensource.org/licenses/EUPL-1.2
  title: Alkemio File Service
  version: 0.2.0
openapi: 3.1.1
paths:
...
  /rest/storage/file/{id}:
    get:
      operationId: PublicHandler.ServeDocument
      parameters:
        - in: path
          name: id
          required: true
          schema:
            format: uuid
            type: string
      responses:
        "200":
          content:
            application/octet-stream:
              schema:
                format: binary
                type: string
          description: OK
        "304":
          description: Not Modified
        "403":
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/http.errorResponse'
          description: Forbidden
        "404":
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/http.errorResponse'
          description: Not Found
        "500":
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/http.errorResponse'
          description: Internal Server Error
        "503":
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/http.errorResponse'
          description: Service Unavailable
      summary: ServeDocument handles GET /rest/storage/file/{id} (and the /rest/storage/document/{id} back-compat alias)
      tags:
        - /rest/storage